requirements.txt
--index-url https://download.pytorch.org/whl/cpu
torch
torchvision
streamlit
librosa
numpy
Pillow